FEM Prescribed displacement working the same way as other FEM constraints
